+++ lib/src/properties/text_field_props.dart
@@ -1,13 +1,12 @@
 import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
+
 import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 class TextFieldProps {
-  TextFieldProps({
-    TextEditingController? controller,
+  const TextFieldProps({
+    this.controller,
     this.decoration,
     this.keyboardType,
     this.textInputAction,
@@ -54,7 +53,7 @@ class TextFieldProps {
     this.scrollPhysics,
     this.autofillHints,
     this.restorationId,
-  }) : this.controller = controller ?? TextEditingController();
+  });
 
   final TextEditingController? controller;